pyproject.toml
[project]
name = "trip-planner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.121.0",
    "httpx>=0.28.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.31",
    "langchain-experimental>=0.3.4",
    "langchain-google-community[places]>=2.0.10",
    "langchain-groq>=0.3.8",
    "langchain-openai>=0.3.35",
    "langchain-tavily>=0.2.13",
    "langgraph>=1.0.1",
    "pydantic>=2.12.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "streamlit>=1.51.0",
    "trulens>=2.4.2",
    "uvicorn>=0.38.0",
]
